+++ BrightWire.Cuda/Helper/DeviceMemory.cs
@@ -31,7 +31,7 @@ namespace BrightWire.Cuda.Helper
             public bool IsValid => true;
 #endif
 
-            public Block(DeviceMemory cache, int index, int size) 
+            public Block(DeviceMemory cache, int index, uint size) 
             {
                 _cache = cache;
                 _index = index;
@@ -88,7 +88,7 @@ namespace BrightWire.Cuda.Helper
 
             public CudaDeviceVariable<float> DeviceVariable => _data;
 	        public CUdeviceptr DevicePointer => _data.DevicePointer;
-	        public int Size => _data.Size;
+	        public uint Size => (uint)_data.Size;
 
 	        public void CopyToDevice(float[] source)
             {
@@ -125,7 +125,7 @@ namespace BrightWire.Cuda.Helper
         readonly int _maxSize;
 		readonly CudaContext _context;
         readonly ConcurrentStack<Layer> _layer = new ConcurrentStack<Layer>();
-        readonly ConcurrentDictionary<int, ThreadSafeHashSet<Block>> _cache = new ConcurrentDictionary<int, ThreadSafeHashSet<Block>>();
+        readonly ConcurrentDictionary<uint, ThreadSafeHashSet<Block>> _cache = new ConcurrentDictionary<uint, ThreadSafeHashSet<Block>>();
         int _index = 0;
 
         public DeviceMemory(CudaContext context, int maxSize)
@@ -199,7 +199,7 @@ namespace BrightWire.Cuda.Helper
             }
         }
 
-        public IDeviceMemoryPtr GetMemory(int size)
+        public IDeviceMemoryPtr GetMemory(uint size)
         {
             Block ret;
