+++ BrightWire/Linear/LogisticRegressionPredictor.cs
@@ -33,7 +33,7 @@ namespace BrightWire.Linear
 
         public float[] Predict(IReadOnlyList<IReadOnlyList<float>> input)
         {
-            using (var feature = _lap.CreateMatrix(input.Count, input[0].Count + 1, (i, j) => j == 0 ? 1 : input[i][j - 1]))
+            using (var feature = _lap.CreateMatrix((uint)input.Count, (uint)input[0].Count + 1, (i, j) => j == 0 ? 1 : input[(int)i][(int)j - 1]))
             using (var h0 = feature.Multiply(_theta))
             using (var h1 = h0.Column(0))
             using (var h = h1.Sigmoid())
