+++ BrightWire/Linear/RegressionPredictor.cs
@@ -24,19 +24,19 @@ namespace BrightWire.Linear
 
         public float Predict(params float[] vals)
         {
-            var v = _lap.CreateVector(vals.Length + 1, i => i == 0 ? 1 : vals[i - 1]);
+            var v = _lap.CreateVector((uint)(vals.Length + 1), i => i == 0 ? 1 : vals[i - 1]);
             return v.DotProduct(_theta);
         }
 
         public float Predict(IReadOnlyList<float> vals)
         {
-            var v = _lap.CreateVector(vals.Count + 1, i => i == 0 ? 1f : vals[i - 1]);
+            var v = _lap.CreateVector((uint)(vals.Count + 1), i => i == 0 ? 1f : vals[(int)(i - 1)]);
             return v.DotProduct(_theta);
         }
 
         public float[] Predict(IReadOnlyList<IReadOnlyList<float>> input)
         {
-            using (var v = _lap.CreateMatrix(input.Count, input[0].Count + 1, (i, j) => j == 0 ? 1 : input[i][j - 1]))
+            using (var v = _lap.CreateMatrix((uint)input.Count, (uint)(input[0].Count + 1), (i, j) => j == 0 ? 1 : input[(int)i][(int)(j - 1)]))
             using (var r = v.Multiply(_theta))
             using(var r2 = r.Row(0)) {
                 return r.AsIndexable().Values.ToArray();
